refac and organize cron and seed scripts
