+++ backend/src/cronJob.ts
@@ -1,29 +1,16 @@
 import cron from "node-cron";
-import { fetchMedia } from "./media";
-import { fetchTextMessagesFromChannels } from "./text";
-import { channels } from "..";
+import { saveMedia, saveTextMessages } from "./client";
 
 cron.schedule("0 * * * *", async () => {
   try {
     console.log("Starting scheduled fetch at", new Date().toISOString());
+    const limit = 50;
 
     // Fetch Memes
-    try {
-      const memeChannelIds = channels.memes;
-      for (const channelId of memeChannelIds) {
-        await fetchMedia(channelId, 10);
-      }
-    } catch (error) {
-      console.error("Error fetching memes:", error);
-    }
+    saveMedia({ limit });
 
     // Fetch Blasts
-    try {
-      const blastChannelIds = channels.blasts;
-      await fetchTextMessagesFromChannels(blastChannelIds, 20);
-    } catch (error) {
-      console.error("Error fetching blasts:", error);
-    }
+    saveTextMessages({ limit });
 
     console.log("Scheduled fetch completed at", new Date().toISOString());
   } catch (error) {
